@@ -22,51 +22,54 @@
   "author": "Dan Halverson <daniel.m.halverson@gmail.com> (http://halversondm.com/)",
   "license": "MIT",
   "dependencies": {
-    "@types/moment-timezone": "0.2.34",
+    "@blueprintjs/datetime": "3.6.0",
+    "@types/moment-timezone": "0.5.10",
     "@types/object-assign": "4.0.30",
-    "@types/react": "15.0.38",
-    "@types/react-dom": "15.5.1",
-    "@types/react-redux": "4.4.46",
-    "@types/react-router": "4.0.14",
-    "@types/requirejs": "2.1.29",
-    "aws-sdk": "2.4.13",
+    "@types/react": "16.7.18",
+    "@types/react-dom": "16.0.11",
+    "@types/react-redux": "6.0.12",
+    "@types/react-router": "4.4.3",
+    "aws-sdk": "2.384.0",
     "bootstrap": "3.3.7",
-    "exceljs": "0.2.24",
+    "exceljs": "1.6.3",
     "font-awesome": "4.7.0",
-    "moment": "2.18.1",
-    "moment-timezone": "0.5.13",
-    "react": "15.6.1",
-    "react-bootstrap": "0.31.1",
-    "react-data-grid": "2.0.53",
-    "react-data-grid-addons": "2.0.54",
-    "react-date-picker": "5.3.28",
-    "react-dom": "15.6.1",
-    "react-redux": "5.0.5",
-    "react-router": "4.1.2",
-    "react-router-dom": "4.1.2",
-    "redux": "3.7.2",
-    "uuid": "3.1.0"
+    "immutable": "4.0.0-rc.12",
+    "moment": "2.23.0",
+    "moment-timezone": "0.5.23",
+    "object-assign": "4.1.1",
+    "react": "16.7.0",
+    "react-bootstrap": "0.32.4",
+    "react-calendar": "2.17.5",
+    "react-data-grid": "5.0.5",
+    "react-data-grid-addons": "5.0.5",
+    "react-dom": "16.7.0",
+    "react-redux": "6.0.0",
+    "react-router": "4.3.1",
+    "react-router-dom": "4.3.1",
+    "redux": "4.0.1",
+    "uuid": "3.3.2"
   },
   "devDependencies": {
-    "awesome-typescript-loader": "3.2.1",
-    "body-parser": "1.17.2",
-    "concurrently": "3.5.0",
-    "copy-webpack-plugin": "4.0.1",
-    "cross-env": "5.0.1",
-    "css-loader": "0.28.4",
-    "express": "4.15.3",
-    "extract-text-webpack-plugin": "2.1.0",
-    "file-loader": "0.11.2",
-    "html-webpack-plugin": "2.29.0",
-    "json-loader": "0.5.6",
-    "morgan": "1.8.2",
-    "redux-devtools": "3.4.0",
-    "rimraf": "2.6.1",
-    "source-map-loader": "0.2.1",
-    "style-loader": "0.18.2",
-    "tslint": "5.5.0",
-    "tslint-react": "3.0.0",
-    "typescript": "2.4.2",
-    "webpack": "2.4.1"
+    "awesome-typescript-loader": "5.2.1",
+    "body-parser": "1.18.3",
+    "concurrently": "4.1.0",
+    "copy-webpack-plugin": "4.6.0",
+    "cross-env": "5.2.0",
+    "css-loader": "2.1.0",
+    "express": "4.16.4",
+    "file-loader": "3.0.1",
+    "html-webpack-plugin": "3.2.0",
+    "json-loader": "0.5.7",
+    "mini-css-extract-plugin": "0.5.0",
+    "morgan": "1.9.1",
+    "redux-devtools": "3.5.0",
+    "rimraf": "2.6.3",
+    "source-map-loader": "0.2.4",
+    "style-loader": "0.23.1",
+    "tslint": "5.12.0",
+    "tslint-react": "3.6.0",
+    "typescript": "3.2.2",
+    "webpack": "4.28.1",
+    "webpack-cli": "3.2.1"
   }
 }
