@@ -1,14 +1,12 @@
 /**
  * Created by Daniel on 7/26/2016.
  */
-import * as moment from "moment";
+import {DateInput, IDateFormatProps} from "@blueprintjs/datetime";
 import * as objectAssign from "object-assign";
 import * as React from "react";
-import { DateField, DatePicker, Footer } from "react-date-picker";
-import "react-date-picker/index.css";
-import { connect } from "react-redux";
-import { TrackerState } from "../reducers/index";
-import { ConnectedState, mapStateToProps} from "./ConnectedState";
+import {connect} from "react-redux";
+import {ConnectedState, mapStateToProps} from "./ConnectedState";
+import moment = require("moment");
 
 class Track extends React.Component<any & ConnectedState & any, any> {
 
@@ -23,11 +21,12 @@ class Track extends React.Component<any & ConnectedState & any, any> {
         this.durationChange = this.durationChange.bind(this);
         this.promptedVisitChange = this.promptedVisitChange.bind(this);
         this.notesChange = this.notesChange.bind(this);
+        this.getMomentFormatter = this.getMomentFormatter.bind(this);
     }
 
     initialState() {
         return {
-            date: moment(),
+            date: new Date(),
             typeOfActivity: "",
             duration: 0,
             typeOfVoid: "",
@@ -39,7 +38,7 @@ class Track extends React.Component<any & ConnectedState & any, any> {
     save(event) {
         event.preventDefault();
         const currentState = this.state;
-        const dataToSend = objectAssign({}, currentState, { profileId: this.props.data.profileId });
+        const dataToSend = objectAssign({}, currentState, {profileId: this.props.data.profileId});
         const data = JSON.stringify(dataToSend);
         const xhr = new XMLHttpRequest();
         xhr.open("POST", "/saveTrack");
@@ -58,29 +57,35 @@ class Track extends React.Component<any & ConnectedState & any, any> {
     }
 
     onDateChange(date) {
-        const newDate = moment.tz(date, "MM/DD/YYYY hh:mm a", moment.tz.guess());
-        const utcDate = newDate.tz("Etc/UTC");
-        this.setState({ date: utcDate });
+        this.setState({date});
     }
 
     typeOfActivityChange(event) {
-        this.setState({ typeOfActivity: event.target.value });
+        this.setState({typeOfActivity: event.target.value});
     }
 
     typeOfVoidChange(event) {
-        this.setState({ typeOfVoid: event.target.value });
+        this.setState({typeOfVoid: event.target.value});
     }
 
     durationChange(event) {
-        this.setState({ duration: event.target.value });
+        this.setState({duration: event.target.value});
     }
 
     promptedVisitChange(event) {
-        this.setState({ promptedVisit: event.target.value });
+        this.setState({promptedVisit: event.target.value});
     }
 
     notesChange(event) {
-        this.setState({ notes: event.target.value });
+        this.setState({notes: event.target.value});
+    }
+
+    getMomentFormatter(format: string): IDateFormatProps {
+        return {
+            formatDate: (date, locale) => moment(date).locale(locale).format(format),
+            parseDate: (str, locale) => moment(str, format).locale(locale).toDate(),
+            placeholder: format,
+        }
     }
 
     render() {
@@ -91,22 +96,9 @@ class Track extends React.Component<any & ConnectedState & any, any> {
                     <div className="form-group">
                         <label htmlFor="date" className="col-sm-2 control-label">Date</label>
                         <div className="col-sm-10">
-                            <DateField forceValidDate={true}
-                                dateFormat="MM/DD/YYYY hh:mm a"
-                                defaultValue={this.state.date}
-                                updateOnDateClick={true}
-                                onChange={this.onDateChange}>
-                                <DatePicker
-                                    navigation={true}
-                                    locale="en"
-                                    forceValidDate={true}
-                                    highlightWeekends={true}
-                                    highlightToday={true}
-                                    weekNumbers={false}
-                                    weekStartDay={0}>
-                                    <Footer todayButton={false} clearButton={false} okButton={false}
-                                        cancelButton={false} /></DatePicker>
-                            </DateField>
+                            <DateInput {...this.getMomentFormatter("MM/DD/YYYY hh:mm a")}
+                                       locale="en" onChange={this.onDateChange} value={this.state.date}
+                                       timePrecision="minute" timePickerProps={{"useAmPm": true}}/>
                         </div>
                     </div>
                     <div className="form-group">
@@ -114,9 +106,9 @@ class Track extends React.Component<any & ConnectedState & any, any> {
                             of Activity</label>
                         <div className="col-sm-10">
                             <select id="typeOfActivity" className="form-control"
-                                onChange={this.typeOfActivityChange}
-                                value={this.state.typeOfActivity}>
-                                <option value="" />
+                                    onChange={this.typeOfActivityChange}
+                                    value={this.state.typeOfActivity}>
+                                <option value=""/>
                                 <option value="Toilet Visit">Toilet Visit</option>
                                 <option value="Underwear Check">Underwear Check</option>
                             </select>
@@ -128,9 +120,9 @@ class Track extends React.Component<any & ConnectedState & any, any> {
                                 or Dry?</label>
                             <div className="col-sm-10">
                                 <select id="typeOfVoid" className="form-control"
-                                    value={this.state.typeOfVoid}
-                                    onChange={this.typeOfVoidChange}>
-                                    <option value="" />
+                                        value={this.state.typeOfVoid}
+                                        onChange={this.typeOfVoidChange}>
+                                    <option value=""/>
                                     <option value="Wet">Wet</option>
                                     <option value="Dry">Dry</option>
                                 </select>
@@ -144,8 +136,8 @@ class Track extends React.Component<any & ConnectedState & any, any> {
                             <div className="col-sm-10">
                                 <div className="input-group">
                                     <input type="number" className="form-control" id="duration"
-                                        value={this.state.duration}
-                                        onChange={this.durationChange} />
+                                           value={this.state.duration}
+                                           onChange={this.durationChange}/>
                                     <div className="input-group-addon">minute(s)</div>
                                 </div>
                             </div>
@@ -156,9 +148,9 @@ class Track extends React.Component<any & ConnectedState & any, any> {
                                 Void</label>
                             <div className="col-sm-10">
                                 <select id="typeOfVoid" className="form-control"
-                                    value={this.state.typeOfVoid}
-                                    onChange={this.typeOfVoidChange}>
-                                    <option value="" />
+                                        value={this.state.typeOfVoid}
+                                        onChange={this.typeOfVoidChange}>
+                                    <option value=""/>
                                     <option value="Urine">Urine</option>
                                     <option value="Bowel Movement">Bowel Movement</option>
                                     <option value="Both">Both</option>
@@ -170,9 +162,9 @@ class Track extends React.Component<any & ConnectedState & any, any> {
                             <label htmlFor="prompted" className="col-sm-2 control-label">Prompted Visit?</label>
                             <div className="col-sm-10">
                                 <select id="prompted" className="form-control"
-                                    value={this.state.promptedVisit}
-                                    onChange={this.promptedVisitChange}>
-                                    <option value="" />
+                                        value={this.state.promptedVisit}
+                                        onChange={this.promptedVisitChange}>
+                                    <option value=""/>
                                     <option value="yes">Yes</option>
                                     <option value="no">No</option>
                                 </select>
@@ -181,11 +173,11 @@ class Track extends React.Component<any & ConnectedState & any, any> {
                     </div>
                     <div className="form-group">
                         <label htmlFor="notes"
-                            className="col-sm-2 control-label">Notes</label>
+                               className="col-sm-2 control-label">Notes</label>
                         <div className="col-sm-10">
                             <textarea className="form-control" id="notes"
-                                value={this.state.notes}
-                                onChange={this.notesChange} />
+                                      value={this.state.notes}
+                                      onChange={this.notesChange}/>
                         </div>
                     </div>
                     <div className="form-group">
